Set up navigation bar only once This allows access to the stack header props in the app bar, fixing issue #136.
